requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
openai
requests # If making HTTP requests to other APIs
python-dotenv # If managing environment variables from a .env file